refactor: extract the app menu test harness into a support file
Keeps app_menu_test.dart under the 500-line max-lines ratchet.
