+++ test/widget/app_menu_test.dart
@@ -1,29 +1,21 @@
-import 'dart:async';
-
-import 'package:alera/src/app/providers.dart';
-import 'package:alera/src/app/theme/alera_dark_theme.dart';
 import 'package:alera/src/core/build_flavor.dart';
 import 'package:alera/src/design_system/feedback/alera_toast.dart';
 import 'package:alera/src/features/app_menu/infra/native_app_menu_channel.dart';
-import 'package:alera/src/features/app_menu/presentation/alera_app_menu_scope.dart';
 import 'package:alera/src/features/app_menu/presentation/app_menu_actions.dart';
-import 'package:alera/src/features/app_window/application/app_window_controller.dart';
-import 'package:alera/src/features/app_window/application/app_window_providers.dart';
-import 'package:alera/src/features/settings/domain/alera_settings.dart';
 import 'package:alera/src/features/updater/domain/alera_update.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:package_info_plus/package_info_plus.dart';
+
+import 'app_menu_test_support.dart';
 
 void main() {
   group('app menu actions', () {
     testWidgets('openAppMenuSettings opens the settings dialog', (
       tester,
     ) async {
-      await _pumpActionHarness(
+      await pumpActionHarness(
         tester,
         onPressed: (context, _) => openAppMenuSettings(context),
       );
@@ -38,8 +30,11 @@ void main() {
     testWidgets('checkForUpdatesFromAppMenu runs check and shows toast', (
       tester,
     ) async {
-      final updateController = _FakeUpdateController(
-        AleraUpdateState(status: AleraUpdateStatus.idle, config: _config()),
+      final updateController = FakeUpdateController(
+        AleraUpdateState(
+          status: AleraUpdateStatus.idle,
+          config: updateConfig(),
+        ),
       );
       final toastMessages = <String>[];
       final sub = AleraToast.stream.listen((data) {
@@ -47,7 +42,7 @@ void main() {
       });
       addTearDown(sub.cancel);
 
-      await _pumpActionHarness(
+      await pumpActionHarness(
         tester,
         updateController: updateController,
         onPressed: (context, ref) => checkForUpdatesFromAppMenu(context, ref),
@@ -62,7 +57,7 @@ void main() {
     });
 
     testWidgets('showAppMenuAbout opens the about dialog', (tester) async {
-      await _pumpActionHarness(
+      await pumpActionHarness(
         tester,
         onPressed: (context, ref) => showAppMenuAbout(
           context,
@@ -89,14 +84,14 @@ void main() {
     testWidgets('about dialog copies the version to the clipboard', (
       tester,
     ) async {
-      _mockClipboard();
+      mockClipboard();
       final toastMessages = <String>[];
       final sub = AleraToast.stream.listen((data) {
         toastMessages.add(data.message);
       });
       addTearDown(sub.cancel);
 
-      await _pumpActionHarness(
+      await pumpActionHarness(
         tester,
         onPressed: (context, ref) => showAppMenuAbout(
           context,
@@ -123,8 +118,11 @@ void main() {
     testWidgets('about dialog runs the update check and closes', (
       tester,
     ) async {
-      final updateController = _FakeUpdateController(
-        AleraUpdateState(status: AleraUpdateStatus.idle, config: _config()),
+      final updateController = FakeUpdateController(
+        AleraUpdateState(
+          status: AleraUpdateStatus.idle,
+          config: updateConfig(),
+        ),
       );
       final toastMessages = <String>[];
       final sub = AleraToast.stream.listen((data) {
@@ -132,7 +130,7 @@ void main() {
       });
       addTearDown(sub.cancel);
 
-      await _pumpActionHarness(
+      await pumpActionHarness(
         tester,
         updateController: updateController,
         onPressed: (context, ref) => showAppMenuAbout(
@@ -158,8 +156,8 @@ void main() {
     });
 
     testWidgets('exitAppFromMenu closes the app window', (tester) async {
-      final window = _FakeAppWindowController();
-      await _pumpActionHarness(
+      final window = FakeAppWindowController();
+      await pumpActionHarness(
         tester,
         window: window,
         onPressed: (_, ref) => exitAppFromMenu(ref),
@@ -179,8 +177,8 @@ void main() {
     ]) {
       group('on $platform (native bridge compatibility)', () {
         testWidgets('renders no in-window menu bar', (tester) async {
-          await _withPlatform(platform, () async {
-            await _pumpMenuScope(tester);
+          await withPlatform(platform, () async {
+            await pumpMenuScope(tester);
 
             expect(find.byType(MenuBar), findsNothing);
             expect(find.byType(PlatformMenuBar), findsNothing);
@@ -188,10 +186,10 @@ void main() {
         });
 
         testWidgets('native channel opens the settings dialog', (tester) async {
-          await _withPlatform(platform, () async {
-            await _pumpMenuScope(tester);
+          await withPlatform(platform, () async {
+            await pumpMenuScope(tester);
 
-            await _invokeNativeMenuMethod(
+            await invokeNativeMenuMethod(
               tester,
               NativeAppMenuMethod.openSettings,
             );
@@ -203,11 +201,11 @@ void main() {
         });
 
         testWidgets('native channel runs the update check', (tester) async {
-          await _withPlatform(platform, () async {
-            final updateController = _FakeUpdateController(
+          await withPlatform(platform, () async {
+            final updateController = FakeUpdateController(
               AleraUpdateState(
                 status: AleraUpdateStatus.idle,
-                config: _config(),
+                config: updateConfig(),
               ),
             );
             final toastMessages = <String>[];
@@ -216,9 +214,9 @@ void main() {
             });
             addTearDown(sub.cancel);
 
-            await _pumpMenuScope(tester, updateController: updateController);
+            await pumpMenuScope(tester, updateController: updateController);
 
-            await _invokeNativeMenuMethod(
+            await invokeNativeMenuMethod(
               tester,
               NativeAppMenuMethod.checkForUpdates,
             );
@@ -231,15 +229,12 @@ void main() {
         });
 
         testWidgets('native channel shows the about dialog', (tester) async {
-          await _withPlatform(platform, () async {
-            _mockPackageInfo(tester);
+          await withPlatform(platform, () async {
+            mockPackageInfo(tester);
 
-            await _pumpMenuScope(tester);
+            await pumpMenuScope(tester);
 
-            await _invokeNativeMenuMethod(
-              tester,
-              NativeAppMenuMethod.showAbout,
-            );
+            await invokeNativeMenuMethod(tester, NativeAppMenuMethod.showAbout);
             await tester.pumpAndSettle();
 
             expect(find.text(kAleraAppName), findsWidgets);
@@ -248,11 +243,11 @@ void main() {
         });
 
         testWidgets('native channel closes the app window', (tester) async {
-          await _withPlatform(platform, () async {
-            final window = _FakeAppWindowController();
-            await _pumpMenuScope(tester, window: window);
+          await withPlatform(platform, () async {
+            final window = FakeAppWindowController();
+            await pumpMenuScope(tester, window: window);
 
-            await _invokeNativeMenuMethod(tester, NativeAppMenuMethod.exitApp);
+            await invokeNativeMenuMethod(tester, NativeAppMenuMethod.exitApp);
             await tester.pump();
 
             expect(window.closeCalls, 1);
@@ -262,11 +257,11 @@ void main() {
         testWidgets(
           'native channel edit methods act on the focused text field',
           (tester) async {
-            await _withPlatform(platform, () async {
+            await withPlatform(platform, () async {
               final controller = TextEditingController();
               addTearDown(controller.dispose);
-              _mockClipboard();
-              await _pumpMenuScope(
+              mockClipboard();
+              await pumpMenuScope(
                 tester,
                 child: Scaffold(
                   body: Center(child: TextField(controller: controller)),
@@ -278,7 +273,7 @@ void main() {
               await tester.enterText(find.byType(TextField), 'hello');
               await tester.pump();
 
-              await _invokeNativeMenuMethod(
+              await invokeNativeMenuMethod(
                 tester,
                 NativeAppMenuMethod.selectAll,
               );
@@ -288,13 +283,13 @@ void main() {
                 const TextSelection(baseOffset: 0, extentOffset: 5),
               );
 
-              await _invokeNativeMenuMethod(tester, NativeAppMenuMethod.cut);
+              await invokeNativeMenuMethod(tester, NativeAppMenuMethod.cut);
               await tester.pump();
               expect(controller.text, isEmpty);
               final clipboard = await Clipboard.getData(Clipboard.kTextPlain);
               expect(clipboard?.text, 'hello');
 
-              await _invokeNativeMenuMethod(tester, NativeAppMenuMethod.paste);
+              await invokeNativeMenuMethod(tester, NativeAppMenuMethod.paste);
               await tester.pump();
               expect(controller.text, 'hello');
             });
@@ -306,240 +301,12 @@ void main() {
     testWidgets('uses PlatformMenuBar on macOS without in-window MenuBar', (
       tester,
     ) async {
-      await _withPlatform(TargetPlatform.macOS, () async {
-        await _pumpMenuScope(tester);
+      await withPlatform(TargetPlatform.macOS, () async {
+        await pumpMenuScope(tester);
 
         expect(find.byType(PlatformMenuBar), findsOneWidget);
         expect(find.byType(MenuBar), findsNothing);
       });
     });
   });
-}
-
-Future<void> _withPlatform(
-  TargetPlatform platform,
-  Future<void> Function() body,
-) async {
-  final previous = debugDefaultTargetPlatformOverride;
-  debugDefaultTargetPlatformOverride = platform;
-  try {
-    await body();
-  } finally {
-    debugDefaultTargetPlatformOverride = previous;
-  }
-}
-
-Future<void> _pumpActionHarness(
-  WidgetTester tester, {
-  required FutureOr<void> Function(BuildContext context, WidgetRef ref)
-  onPressed,
-  _FakeUpdateController? updateController,
-  _FakeAppWindowController? window,
-}) async {
-  await tester.pumpWidget(
-    ProviderScope(
-      overrides: [
-        aleraUpdateControllerProvider.overrideWith(
-          () =>
-              updateController ??
-              _FakeUpdateController(
-                AleraUpdateState(
-                  status: AleraUpdateStatus.idle,
-                  config: _config(),
-                ),
-              ),
-        ),
-        appWindowControllerProvider.overrideWithValue(
-          window ?? _FakeAppWindowController(),
-        ),
-        settingsControllerProvider.overrideWith(
-          () => _FakeSettingsController(AleraSettings.defaults),
-        ),
-      ],
-      child: MaterialApp(
-        theme: buildAleraDarkTheme(),
-        home: Scaffold(
-          body: Consumer(
-            builder: (context, ref, _) {
-              return TextButton(
-                onPressed: () =>
-                    unawaited(Future<void>.sync(() => onPressed(context, ref))),
-                child: const Text('Run'),
-              );
-            },
-          ),
-        ),
-      ),
-    ),
-  );
-}
-
-Future<void> _pumpMenuScope(
-  WidgetTester tester, {
-  _FakeUpdateController? updateController,
-  _FakeAppWindowController? window,
-  Widget? child,
-}) async {
-  await tester.pumpWidget(
-    ProviderScope(
-      overrides: [
-        aleraUpdateControllerProvider.overrideWith(
-          () =>
-              updateController ??
-              _FakeUpdateController(
-                AleraUpdateState(
-                  status: AleraUpdateStatus.idle,
-                  config: _config(),
-                ),
-              ),
-        ),
-        appWindowControllerProvider.overrideWithValue(
-          window ?? _FakeAppWindowController(),
-        ),
-        settingsControllerProvider.overrideWith(
-          () => _FakeSettingsController(AleraSettings.defaults),
-        ),
-      ],
-      child: MaterialApp(
-        theme: buildAleraDarkTheme(),
-        home: AleraAppMenuScope(
-          child: child ?? const Scaffold(body: SizedBox.expand()),
-        ),
-      ),
-    ),
-  );
-  await tester.pump();
-}
-
-/// Simulates a native (GTK/Win32) menu activation arriving over the app-menu
-/// channel.
-Future<void> _invokeNativeMenuMethod(WidgetTester tester, String method) async {
-  await TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-      .handlePlatformMessage(
-        nativeAppMenuChannelName,
-        const StandardMethodCodec().encodeMethodCall(MethodCall(method)),
-        (_) {},
-      );
-}
-
-void _mockPackageInfo(WidgetTester tester) {
-  const channel = MethodChannel('dev.fluttercommunity.plus/package_info');
-  TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-      .setMockMethodCallHandler(channel, (call) async {
-        return <String, String>{
-          'appName': kAleraAppName,
-          'packageName': 'dev.leynier.alera',
-          'version': '1.2.3',
-          'buildNumber': '45',
-        };
-      });
-  addTearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, null);
-  });
-}
-
-/// The test binding does not answer clipboard platform messages, so the edit
-/// menu intents would hang without an in-memory mock.
-void _mockClipboard() {
-  String? clipboardText;
-  final messenger =
-      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger;
-  messenger.setMockMethodCallHandler(SystemChannels.platform, (call) async {
-    switch (call.method) {
-      case 'Clipboard.getData':
-        final text = clipboardText;
-        return text == null ? null : <String, dynamic>{'text': text};
-      case 'Clipboard.setData':
-        clipboardText =
-            (call.arguments as Map<Object?, Object?>)['text'] as String?;
-    }
-    return null;
-  });
-  addTearDown(() {
-    messenger.setMockMethodCallHandler(SystemChannels.platform, null);
-  });
-}
-
-AleraUpdateConfig _config() {
-  return AleraUpdateConfig(
-    archiveUrl: Uri.parse('https://example.com/app-archive.json'),
-    releasePageUrl: Uri.parse('https://example.com/releases'),
-    channel: AleraUpdateChannel.stable,
-    autoInstallEnabled: true,
-    signedRelease: true,
-  );
-}
-
-class _FakeSettingsController extends SettingsController {
-  _FakeSettingsController(this._seed);
-
-  final AleraSettings _seed;
-
-  @override
-  AleraSettings build() => _seed;
-}
-
-class _FakeUpdateController extends AleraUpdateController {
-  _FakeUpdateController(this._seed);
-
-  final AleraUpdateState _seed;
-  int checkForUpdatesCalls = 0;
-
-  @override
-  AleraUpdateState build() => _seed;
-
-  @override
-  Future<void> checkForUpdates() async {
-    checkForUpdatesCalls += 1;
-    state = state.copyWith(
-      status: AleraUpdateStatus.notAvailable,
-      message: 'Alera is up to date.',
-    );
-  }
-}
-
-class _FakeAppWindowController implements AppWindowController {
-  int closeCalls = 0;
-
-  @override
-  void addListener(AppWindowEventListener listener) {}
-
-  @override
-  Future<void> close() async {
-    closeCalls += 1;
-  }
-
-  @override
-  Future<void> destroy() async {}
-
-  @override
-  Future<Rect> getBounds() async => const Rect.fromLTWH(0, 0, 1280, 720);
-
-  @override
-  Future<bool> isFullScreen() async => false;
-
-  @override
-  Future<bool> isMaximized() async => false;
-
-  @override
-  Future<bool> isMinimized() async => false;
-
-  @override
-  Future<void> maximize() async {}
-
-  @override
-  void removeListener(AppWindowEventListener listener) {}
-
-  @override
-  Future<void> setBounds(Rect bounds) async {}
-
-  @override
-  Future<void> setFullScreen(bool value) async {}
-
-  @override
-  Future<void> setPreventClose(bool value) async {}
-
-  @override
-  Future<void> setTitle(String title) async {}
 }
